Member: Join chain and send heartbeat to master
Updates `MEMBER.CONNECT_TO_MASTER` so that it implements the following:

1. Send a heartbeat to the master.
    The heartbeat is implemented with an etcd Lease, where a client
requests a timeout and receives a session token that can be used to send
KeepAlives. If a KeepAlive is not received within the timeout, the lease
expires. The token can also be associated with keys and locks, with
the result that the key or lock is deleted if the session times out.

2. Acquire a unique ID for the member (unique only among the chain).
    The heartbeat makes it possible to use a distributed lock without
the risk of leaking the lock when a member crashes. This is used to
implement atomic incrementing of a key value in etcd that is the source
of unique IDs for chain members.

3. Store arbitrary info associated with own ID in etcd.
    Each member PUTs a key in order to claim membership in a chain:
    ```
    PUT credis/$chain_id/$member_id "<anything here>"
    ```
The heartbeat lease from (1) is used to specify that this key should be
deleted if the heartbeat times out. By watching changes to this key, the
master service can tell if the member is alive, and receive
communication from the member, such as connection info and metrics.
I'm still defining the exact interface that specifies how this is used.