Member: Watch own config and process updates
Updates the member module so that it implements the following:

1. Write own connection info and current config in etcd.
    When connected to a cluster, the member now reports Redis
connection info and its current prev/next/role. When these configuration
values change, they are updated in etcd.

2. Watch configs pushed by the master and update accordingly.
    The member listens to a stream of changes to its own intended
configuration, which the master writes to etcd. When a change happens,
it sends the appropriate invocation of MEMBER.SET_ROLE to _itself_ in
order to make the change happen.

3. New parent blocks until its new child has reported updated config
    As discussed in the CR paper, the parent must begin to forward
writes _after_ the child begins to forward acks. To ensure this, a new
parent waits until it sees itself reported as its child's parent. See
design doc and CR page 5 under "Failure of other servers." As the paper
calls it, this is maintaining the "Update Propagation Invariant."