Refactor etcd client to use I/O params
Previously the client methods returned smart pointers to responses.
Instead, they now take a trailing output parameter.

In general, the API has been significantly simplified to comply with the
Google style guide section on argument passing, which suggests that an
options struct (in this case, Request object) is generally better than a
bunch of parameters with some defaults.