Implement fault-tolerant master
This implements a master module that keeps all its state in etcd, so
that it can be killed and restarted without disruption to the chain.

It responds to the same commands as the current master, except that
`MASTER.REMOVE` is fully implemented.

It detects failures by watching for heartbeat keys in etcd, which are
published with a chain-specific prefix and are just the connection info
of the relevant node. Members publish this key when processing the
`MEMBER.CONNECT_TO_MASTER` command, and an aeTimeEvent on their event
loop refreshes the TTL of the key on a configurable interval.

However, the client can notify the master that a node died by calling
`MASTER.REMOVE` with the relevant connection info too.