Fault-tolerant failure-detecting etcd-based master
This implements etcd master support in credis.

Break this up and stuff!

The master itself is also a redis server so that it can support the
REFRESH_HEAD/REFRESH_TAIL etc. API. However, it would be really slow to
use as a regular KVS because it spends a lot of time watching the state
of the chain in etcd and writing intended configuration to the members.

I'll break this commit down soon.