Implement etcd master client
This commit implements the etcd master client, which is different from
the current master client in  that it connects to etcd instead of to
another Redis server.

On connecting, it writes its connection info to a key with a TTL that
must be periodically refreshed. If the TTL expires, the key disappears.
This way the master can tell if a node has died by watching changes to
all such heartbeat keys.

The head and tail read and write checkpoints directly into etcd.