Bugfixes after porting zh's add-node handling
One change is that we reconnect the read_context to the event loop after
reconnecting ACK context. Not doing this resulted in dropped GET replies
and ultimately hanging the benchmark because nothing kicks off the next
request.

The weird change is reordering the order that connections are
reconnected in, on line 168. This fixes an issue where the actual
*memory addresses* of the contexts change between reconnects, which
causes some weirdness in how responses are routed and ultimately leads
to accessing a null pointer when reading what the program believes is a
reply to a GET. Specifically, the values of the `client.read_context_`
and `client.ack_subscribe_context_` pointers are switched when
reconnecting after a tail failure.

I have literally no idea why the memory addresses of the contexts
matter, only a hunch as to why this fixes it, and no idea whether the
change is portable. That's why this branch is called the dirty one.